feat: construct.node
Introduce `construct.node` as an alternative to `Node.of(construct)` in order to be compatible with the AWS CDK.

This means that we no longer need a way to customize the construct node factory, so the relevant types were also removed.

BREAKING CHANGE: `Node.of(construct)` was replaced with `construct.node`.