feat: remove onPrepare() and onSynthesize()
We realized that synthesis (and prepare) is a domain-specific activity and therefore both in AWS CDK and cdk8s it is implemented centrally at the app level.

Introduce `lock()` and `unlock()` as public APIs to enable domains to lock the tree during their synthesis operations.

BREAKING CHANGE: constructs no longer have an `onPrepare()` and `onSynthesis()` hooks. See domain-specific (such as AWS CDK and cdk8s) for alternatives.