feat: introduce IValidation instead of onValidate()
Instead of a protected method called `onValidate()` which pollutes the namespace for all constructs, introduce an interface `IValidation` which contains a `validate(): string[]` method.

The `construct.node.validate()` method can be used to validate any construct. If it implements `IValidation`, it's `validate()` method will be called. Otherwise, an empty list will be returned.

BREAKING CHANGE: to add validation logic for constructs, implement the `IValidation` interface. 
* The `construct.node.validate()` method no longer validates the entire tree, but only the specific construct on which it is called.