feat: addMetadata() no longer attaches stack traces by default
Stack traces are only attached if `addMetadata()` is called with the option `stackTrace: true`.

`addInfo()`, `addWarning()`, and `addError()` also attach stack traces.

BREAKING CHANGE: Stack traces are not attached to metadata entries by default. To include stack trace in a metadata entry, pass `{ stackTrace: true }` as the 3rd parameter, which is now of type `MetadataOptions` and not a `fromFunction` pointer.