chore: restore to version 3.2.78
Due to a bug in the release workflow, the bump task did not have tags
in order to be identify the changes between versions and since weve
had breaking changes in the past, we accidentally had a few unintentional
major version bumps.

This commit restores the version to 3.2.78, which was the last version
before the incident.