chore: complete JSII 5.x upgrade (#1978)
The previous related change #1630 upgraded just the `jsii` library, however, the biggest advantage of 5.x is to stay in sync with the Typescript version but that was kept as 4.x.
In addition, Jest was kept as a very old 27 version (due to projen logic with pre-existing version) while it should be upgraded to the current version.
This PR fixes that problems.
Also updates some related dependencies